Extract test policies from Pundit integration spec
